Update tests and docs for new detect_lang func
