+++ docs/source/getting_started/quickstart.rst
@@ -453,12 +453,12 @@ For example:
     >>> en_corpus = textacy.Corpus(
     ...     "en", data=(
     ...         text for text in texts
-    ...         if textacy.text_utils.detect_language(text) == "en")
+    ...         if textacy.lang_utils.detect_lang(text) == "en")
     ... )
     >>> es_corpus = textacy.Corpus(
     ...     "es", data=(
     ...         text for text in texts
-    ...         if textacy.text_utils.detect_language(text) == "es")
+    ...         if textacy.lang_utils.detect_lang(text) == "es")
     ... )
 
 Both of these options are less convenient than I'd like, but hopefully they
