+++ tests/test_text_utils.py
@@ -51,15 +51,6 @@ CRUFTY_TERMS = [
 GOOD_TERMS = ["foo (bar)", "foo?", "bar!", "-123.4"]
 BAD_TERMS = ["(foo bar", "foo) bar", "?>,!-.", "", "foo) (bar"]
 
-LANG_SENTS = [
-    ("en", "This sentence is in English."),
-    ("es", "Esta oración es en Español."),
-    ("fr", "Cette phrase est en français."),
-    ("un", "1"),
-    ("un", " "),
-    ("un", ""),
-]
-
 TEXT = """
     The hedge fund magnates Daniel S. Loeb, Louis Moore Bacon and Steven A. Cohen have much in common. They have managed billions of dollars in capital, earning vast fortunes. They have invested millions in art — and millions more in political candidates.
     Moreover, each has exploited an esoteric tax loophole that saved them millions in taxes. The trick? Route the money to Bermuda and back.
@@ -90,11 +81,6 @@ def test_is_acronym_bad():
 
 def test_is_acronym_exclude():
     assert not text_utils.is_acronym("NASA", exclude={"NASA"})
-
-
-def test_detect_language():
-    for lang, sent in LANG_SENTS:
-        assert text_utils.detect_language(sent) == lang
 
 
 def test_keyword_in_context_keyword():
